Assorted changes to the matrix print functions.
* imageadjust.cpp (ia_print_matrix): Add an argument that allows us to define
the ending char, this comes in handy when printing for logs or for output.  Also
make sure that the arguments are const.
(ia_print_matrix_vector): Make sure we flush the stdout.
(main): Stop trying to print the matrices at the end of the execution.
(ia_calculate_and_capture): Change the printing scheme for the rvector and
tvector.  We now print them both in one line.
Change the delay time to 250 ms.