Start moving to C++ only.
There was an issue were I could not run the application compiled with
opencv2.2.0.  I followed the issue to the use of a pointer.  After thinking
about it for a bit, I decided to move all the code to only C++.  This is the
first part of the move.  There are still parts of the code that contain C
specific stuff.  But the code is at a point where it compiles and runs.
These are the files that changed:
* IA_Square.cpp (calculate_rgb, IA_ChessboardImage, debug_print):
* IA_Square.h (IA_ChessboardImage):
* ia_core.cpp (ia_calculate_and_capture)
(ia_information_extraction_debug, ia_image_calc_intr)
(ia_imageadjust, ia_video_calc_intr):
* ia_core.h:
* ia_input.cpp (ia_init_input_struct, ia_print_input_struct)
(ia_get_intrinsics_from_file, ia_create_config, ia_init_input)
(ia_free_input_struct):
* ia_input.h (iif, vid_file, images, objective, checked):
* imageadjust.cpp (main):