Optimize the horizontal vertical tests.
* IA_Square.cpp (IA_Line): Implement a binary var that will contain information
of the type of line  (horizontal, vertical or normal).
(resolve_width, resolve_height): Test for normal (which is the most common case)
in the beginning of the function.  We can then test for the other types.
* IA_Square.h (line_t, VERTICAL, HORIZONTAL, NORMAL, l_type)
(horizontal, vertical): Implement a new enumeration that will contain the line
types: HORIZONTAL, VERTICAL and NORMAL.  Also add a binary variable l_type of
type line_t that will carry the line type information.