Have calc_rgb calculate accum based on an rgb table.
We are moving towards this (more complex) way of calculating the accum offset
because it facilitates analyzing the color range around 0,256.
We have a table of 8 positions.  Position 0 should always be 0.  The positions
represent the maximum values of the range of a color   So position 1 contains 21
which is the max position for the yellow color.  In this commit, this is
true for every color except red.  Red is special because it contains the
beginning and the end of the spectrum.
We select an accum offset when we find a number in the range table that is
greater than the pixel we are currently on.  The offset we select is the one
that is previous to this greater number (i-1).
Note that this is not how things are going to look.  Once we have implemented
the dynamic range calculation, we will use that created range as the argument
for calculate_rgb.
* IA_Square.cpp (IA_Square, calculate_rgb, IA_ChessboardImage): Introduce the
range table to the calc_rgb call.
* IA_Square.h (IA_Square, sqr): Likewise.