Be stricter in the type of chessboard we accept.
We need to be able to rely on the point of reference for our calculations.
A referenced square needs to be the same regardless of chessboard orientation.
The only way we can ensure this is by enforcing chessboards with odd
dimensions. Refer to "Learning Opencv page 382.
* ia_input.cpp (ia_usage, ia_init_input): Modify the user documentation to
reflect the new argument setup.  Enforce the relationship between the chessboard
dimensions.
* ia_input.h (b_size): Add some additional comments.