Completely control the rounding of square dimensions.
For the dimensions we want the origin (top-left corner) to be rounded down.
Additionally, we want the sizes (height and width) to be rounded up.  We do this
to avoid having a square of size 0 when the height and the width are numbers
between 0 and 1.
* IA_Square.cpp (IA_Square, IA_ChessboardImage):
* IA_Square.h (FLOOR_MIN, CEIL_DIST): New defines that replace the previous
O_S*, X* and Y*.  This makes the code shorter :).
(MIN_4, O_S_HEIGHT, O_S_WIDTH, X_MAX, X_MIN, Y_MAX, Y_MIN):  Removed defines.