Separate id calculation logic.
We also created a new error.h file to put all the errors.
* CMakeLists.txt:  Consider ilacLabeler.{h,cpp} in the makefiles.
* error.h: Put all the error logic from ilacSquare.h into error.h.
* ilacLabeler.cpp: New file that holds the id calculation logic.  It
has all the relevant functions from ilacSquare that did this.  It now holds
ILAC_Square class.  We need to think of another name for ilacSquare.{cpp,h}
files.
* ilacLabeler.h: New header for ilacLabeler.cpp.
* ilacSquare.cpp: We removed all the ILAC_Square logic from this file.  We also
removed the logic needed to calculate the chessboard id.
* ilacSquare.h: Likewise. Removed all the Exception classes.