Add logic for detecting colored spheres.
* include/ilacLabeler.h (ILAC_Sphere, ILAC_SphereFinder): New classes.
(img, center, radius): Variables that represent the sphere. A sphere on an image
will be a circle.
* src/ilacLabeler.cpp (ILAC_Sphere, ILAC_SphereFinder): New classes for
handling the spheres.
(findSpheres): Method that will detect the spheres inside an image. Returns a
vector of ILAC_Sphere objects organized in order of radius size. With the
largest radius at the beginning.