Add real object measurements.
We will use the real dimensions of the sphere and the chessboard square to help
detect the spheres. No specific units need to be used. The only condition is for
both the square size and the sphere diameter to be in the same units.
* include/ilacImage.h (ILAC_Image): Add new variables to this class.
pixPerUU, sphDiamUU, sqrSideUU): New class variables that hold the pixel per
Unit of measurement, sphere diameter and square side size respectively
* src/_ilac.cpp (IlacCB_init): Use the new ILAC_Image constructor.
* src/ilacImage.cpp (ILAC_Image, calcPixPerUU): New helper function that
calculates the pixel per Unit of measurement.