Remove the sphere colored rectangle from mask.
When we do the mask the rectangle that contains the color of the spheres
persists. We should remove it as we know exactly were it is.
* include/ilacLabeler.h (enc_rect): Add new opencv rectangle to hold the
coordinates of the enclosing rectangle for all the squares.
* src/ilacLabeler.cpp (ILAC_Square): Initialized enc_rect when we create the
enclosing rectangle.
(findSpheres): Use opencv's rectangle to paint a black rectangle on top of the
chessboard square containing the sphere color.