Be smarter at removing the noise.
Implemented a morphological process that incrementally reduces the noise by
increasing the size of the structuring element.
* include/error.h (ILACExBadMaskProcess): New error method. Thrown when we
cannot eliminate noise.
* src/ilacLabeler.cpp (findSpheres): Incrementally increase the open
structuring elment size until the difference between increments is less
than a threshold (for {col,row}Ratio). As a heuristic {col,row}Ratio<15%,
which means that the physical spheres need to be smallish :)